Also move tests for staking_test.go
